PKMApi/requirements.txt
asgiref==3.8.1
boto3==1.35.40
botocore==1.35.40
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-dotenv==1.4.2
django-polymorphic==3.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
setuptools==75.6.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
